requirements.txt
python-dotenv
pandas
llama-index
ollama
flask
flask-cors
werkzeug
jinja2
bs4